requirements.txt
torch
torchvision
efficientnet_pytorch
pillow 
